# openapi.json
{
  "components": {
    "schemas": {
...
      "CreateScheduleRequest": {
        "properties": {
          "job_duration_minutes": {
            "default": 30,
            "maximum": 1440.0,
            "minimum": 1.0,
            "title": "Job Duration Minutes",
            "type": "integer"
          },
          "max_delay_hours": {
            "default": 24,
            "maximum": 168.0,
            "minimum": 1.0,
            "title": "Max Delay Hours",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "org_id": {
            "title": "Org Id",
            "type": "string"
          },
          "preferred_regions": {
            "items": {
              "type": "string"
            },
            "title": "Preferred Regions",
            "type": "array"
          },
          "providers": {
            "items": {
              "type": "string"
            },
            "title": "Providers",
            "type": "array"
          },
          "strategy": {
            "$ref": "#/components/schemas/ScheduleStrategy",
            "default": "lowest_carbon"
          }
        },
        "required": [
          "name",
          "org_id"
        ],
        "title": "CreateScheduleRequest",
        "type": "object"
      },
      "CronSchedule": {
        "description": "A recurring job with carbon-aware scheduling.",
        "properties": {
          "active": {
            "default": true,
            "title": "Active",
            "type": "boolean"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "job_duration_minutes": {
            "maximum": 1440.0,
            "minimum": 1.0,
            "title": "Job Duration Minutes",
            "type": "integer"
          },
          "max_delay_hours": {
            "default": 24,
            "description": "Max hours to delay for green window",
            "maximum": 168.0,
            "minimum": 1.0,
            "title": "Max Delay Hours",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "org_id": {
            "title": "Org Id",
            "type": "string"
          },
          "preferred_regions": {
            "items": {
              "type": "string"
            },
            "title": "Preferred Regions",
            "type": "array"
          },
          "providers": {
            "items": {
              "type": "string"
            },
            "title": "Providers",
            "type": "array"
          },
          "strategy": {
            "$ref": "#/components/schemas/ScheduleStrategy",
            "default": "lowest_carbon"
          }
        },
        "required": [
          "id",
          "name",
          "org_id",
          "job_duration_minutes",
          "created_at"
        ],
        "title": "CronSchedule",
        "type": "object"
      },
...
      "FindWindowRequest": {
        "properties": {
          "job_duration_minutes": {
            "default": 30,
            "maximum": 1440.0,
            "minimum": 1.0,
            "title": "Job Duration Minutes",
            "type": "integer"
          },
          "max_delay_hours": {
            "default": 24,
            "maximum": 168.0,
            "minimum": 1.0,
            "title": "Max Delay Hours",
            "type": "integer"
          },
          "preferred_regions": {
            "items": {
              "type": "string"
            },
            "title": "Preferred Regions",
            "type": "array"
          },
          "providers": {
            "items": {
              "type": "string"
            },
            "title": "Providers",
            "type": "array"
          },
          "strategy": {
            "$ref": "#/components/schemas/ScheduleStrategy",
            "default": "lowest_carbon"
          }
        },
        "title": "FindWindowRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ScheduleRecommendation": {
        "description": "Recommendation for when and where to run a job.",
        "properties": {
          "alternatives": {
            "items": {
              "$ref": "#/components/schemas/TimeSlot"
            },
            "title": "Alternatives",
            "type": "array"
          },
          "carbon_saved_vs_now_pct": {
            "description": "Percentage of carbon saved vs running immediately",
            "title": "Carbon Saved Vs Now Pct",
            "type": "number"
          },
          "evaluated_slots": {
            "title": "Evaluated Slots",
            "type": "integer"
          },
          "forecast": {
            "description": "The recommended region's hourly intensity curve across the window",
            "items": {
              "$ref": "#/components/schemas/TimeSlot"
            },
            "title": "Forecast",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "job_duration_minutes": {
            "title": "Job Duration Minutes",
            "type": "integer"
          },
          "marginal_basis": {
            "default": "heuristic",
            "description": "Whether the recommended region's marginal signal is 'measured' (from a configured marginal source) or 'heuristic' (a merit-order estimate)",
            "title": "Marginal Basis",
            "type": "string"
          },
          "recommended": {
            "$ref": "#/components/schemas/TimeSlot"
          },
          "strategy": {
            "$ref": "#/components/schemas/ScheduleStrategy"
          },
          "window_end": {
            "format": "date-time",
            "title": "Window End",
            "type": "string"
          },
          "window_start": {
            "format": "date-time",
            "title": "Window Start",
            "type": "string"
          }
        },
        "required": [
          "id",
          "recommended",
          "alternatives",
          "job_duration_minutes",
          "window_start",
          "window_end",
          "strategy",
          "carbon_saved_vs_now_pct",
          "evaluated_slots"
        ],
        "title": "ScheduleRecommendation",
        "type": "object"
      },
      "ScheduleStrategy": {
        "description": "How to optimize the schedule.",
        "enum": [
          "lowest_carbon",
          "highest_renewable",
          "balanced"
        ],
        "title": "ScheduleStrategy",
        "type": "string"
      },
...
      "TimeSlot": {
        "description": "A candidate time window for scheduling.",
        "properties": {
          "carbon_intensity_gco2_kwh": {
            "title": "Carbon Intensity Gco2 Kwh",
            "type": "number"
          },
          "clean_surplus": {
            "default": false,
            "description": "True when this slot looks like clean oversupply -- renewables dominant, very low carbon -- so it's the highest-value time to run (near-zero marginal). Given a bounded ranking edge. A heuristic, not measured curtailment.",
            "title": "Clean Surplus",
            "type": "boolean"
          },
          "end": {
            "format": "date-time",
            "title": "End",
            "type": "string"
          },
          "grid_zone": {
            "title": "Grid Zone",
            "type": "string"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "region": {
            "title": "Region",
            "type": "string"
          },
          "renewable_percentage": {
            "title": "Renewable Percentage",
            "type": "number"
          },
          "score": {
            "description": "Lower is better for carbon, higher for renewable",
            "title": "Score",
            "type": "number"
          },
          "start": {
            "format": "date-time",
            "title": "Start",
            "type": "string"
          }
        },
        "required": [
          "start",
          "end",
          "provider",
          "region",
          "grid_zone",
          "carbon_intensity_gco2_kwh",
          "renewable_percentage",
          "score"
        ],
        "title": "TimeSlot",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "A free, open API for the real-time carbon intensity of cloud regions. No account or API key required; usage is rate-limited so it stays responsive for everyone.\n\n## Carbon Data API\n- **8 live grid-operator integrations** (UK, EIA, OpenElectricity/AEMO, IESO/AESO, Taipower, GridStatus, ENTSO-E, Electricity Maps) plus labeled heuristic and weather-based estimates, cascading\n- **Grid carbon intensity** for 75+ cloud regions; every response is tagged with its `source`\n- **Batch queries** for multiple regions in a single call\n\n## Compliance Reporting\n- **GHG-Protocol-structured** Scope 2 (location-based) + Scope 3 Cat 1 reporting\n- Drafts aimed at **CSRD / SEC Climate / SB 253** workflows (not an assured report)\n- **EU Taxonomy** eligibility flag (simplified screening)\n\n## Green SLA Monitoring (Beta)\n- **Define carbon targets** \u2014 max gCO2/kWh, min renewable %\n- **On-demand + background checks** (in-memory state)\n- **Attestation-style summary reports**\n\n## Carbon-Aware Scheduling\n- **Find optimal time windows** for batch jobs, CI/CD, ML training\n- **Multi-region evaluation** across AWS, GCP, Azure\n- **Three strategies** \u2014 lowest carbon, highest renewable, balanced\n- Returns a recommendation \u2014 it doesn't run or defer your workload itself",
    "title": "Carbon Lens",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/scheduler/find-window": {
      "post": {
        "description": "Find the optimal low-carbon time window for a batch job.\n\nEvaluates carbon intensity across all regions and time slots within the\nspecified window, and returns the greenest option.\n\nUse this to decide when to run CI/CD pipelines, batch processing,\nML training, or any deferrable workload.",
        "operationId": "find_optimal_window_api_v1_scheduler_find_window_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FindWindowRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleRecommendation"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Find Optimal Window",
        "tags": [
          "Scheduling"
        ]
      }
    },
    "/api/v1/scheduler/now": {
      "get": {
        "description": "Quick check: what's the greenest region to run a job RIGHT NOW?\n\nSimpler than /find-window \u2014 just tells you the best region at this moment.",
        "operationId": "best_region_now_api_v1_scheduler_now_get",
        "parameters": [
          {
            "in": "query",
            "name": "duration_minutes",
            "required": false,
            "schema": {
              "default": 30,
              "maximum": 1440,
              "minimum": 1,
              "title": "Duration Minutes",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "providers",
            "required": false,
            "schema": {
              "default": "aws,gcp,azure",
              "title": "Providers",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleRecommendation"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Best Region Now",
        "tags": [
          "Scheduling"
        ]
      }
    },
    "/api/v1/scheduler/schedules": {
      "get": {
        "description": "List all scheduled jobs for an organization.",
        "operationId": "list_schedules_api_v1_scheduler_schedules_get",
        "parameters": [
          {
            "in": "query",
            "name": "org_id",
            "required": true,
            "schema": {
              "title": "Org Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CronSchedule"
                  },
                  "title": "Response List Schedules Api V1 Scheduler Schedules Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "List Schedules",
        "tags": [
          "Scheduling"
        ]
      },
      "post": {
        "description": "Create a recurring carbon-aware schedule.\n\nDefine a job's requirements and the scheduler will recommend optimal\nexecution windows each time it runs.",
        "operationId": "create_schedule_api_v1_scheduler_schedules_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateScheduleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CronSchedule"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Create Schedule",
        "tags": [
          "Scheduling"
        ]
      }
    },
    "/api/v1/scheduler/schedules/{schedule_id}": {
      "delete": {
        "description": "Delete a schedule.",
        "operationId": "delete_schedule_api_v1_scheduler_schedules__schedule_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "schedule_id",
            "required": true,
            "schema": {
              "title": "Schedule Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Delete Schedule Api V1 Scheduler Schedules  Schedule Id  Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Delete Schedule",
        "tags": [
          "Scheduling"
        ]
      },
      "get": {
        "description": "Get a specific schedule.",
        "operationId": "get_schedule_api_v1_scheduler_schedules__schedule_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "schedule_id",
            "required": true,
            "schema": {
              "title": "Schedule Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CronSchedule"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Get Schedule",
        "tags": [
          "Scheduling"
        ]
      }
    },
    "/api/v1/scheduler/schedules/{schedule_id}/next": {
      "post": {
        "description": "Get the next optimal execution window for a scheduled job.\n\nUses the schedule's configuration to find the best upcoming time slot.",
        "operationId": "get_next_window_api_v1_scheduler_schedules__schedule_id__next_post",
        "parameters": [
          {
            "in": "path",
            "name": "schedule_id",
            "required": true,
            "schema": {
              "title": "Schedule Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleRecommendation"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Get Next Window",
        "tags": [
          "Scheduling"
        ]
      }
    },
...
  },
...
}